% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coda_df.R
\name{coda_df}
\alias{coda_df}
\title{Create a dataframe from an mcmc.list object}
\usage{
coda_df(coda.object, parameters = NULL)
}
\arguments{
\item{coda.object}{an mcmc.list object}

\item{parameters}{character vector of parameter names to include in dataframe. If none are
     supplied all monitored parameters are included.}
}
\value{
A dataframe
}
\description{
\code{coda_df} takes as input an mcmc.list object and converts it to a
     dataframe with a column for each monitored parameter and two indicator
     columns, one giving the chain number and another giving the iteration
     number within that chain.
}
\examples{
library(coda)
data(line)

df <- coda_df(line)
head(df)
tail(df)

df <- coda_df(line, parameters = "alpha")
head(df)

df <- coda_df(line, parameters = c("alpha", "beta"))
head(df)

df <- coda_df(line, parameters = grep("sig", coda::varnames(line), value = TRUE))
head(df)

df <- coda_df(line, parameters = grep("a", coda::varnames(line), value = TRUE))
head(df)

df <- coda_df(line, parameters = c("alpha", grep("sig", coda::varnames(line),
              value = TRUE)))
head(df)
}
\author{
Michael Malick
}
\seealso{
\code{\link{data.frame}} \code{\link{mcmc}} \code{\link{mcmc.list}}
}

