% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coda_diag.R
\name{coda_diag}
\alias{coda_diag}
\title{Create multiple diagnostic graphics for a single parameter}
\usage{
coda_diag(coda.object, parameters = NULL)
}
\arguments{
\item{coda.object}{an mcmc.list object}

\item{parameters}{character vector of parameter names to create diagnostic plots for. If
     none are supplied all monitored parameters are included.}
}
\value{
A graphic with six diagnostic plots:

    \itemize{
        \item{traceplot of all chains}
        \item{density plot of marginal distribution for each chain}
        \item{autocorrelation for each chain with effective sample size
              (ess)}
        \item{evolution of the Gelman-Rubin statistic with 97.5\% credibility
              interval}
        \item{evolution of the median value for the parameter with 95\%
              credibility interval}
        \item{histogram of the marginal distribution of all chains
              combined with median value and 95\% credibility interval}
     }
}
\description{
\code{coda_diag} creates a graphic that includes six diagnostic plots to
     help monitor convergence of an MCMC chain for a parameter. By default
     the function will loop through all parameters in an mcmc.list object and
     create a single graphic of six diagnostic plots for each parameter.
}
\examples{
library(coda)
data(line)

coda_diag(line)

coda_diag(line, parameters = "alpha")

coda_diag(line, parameters = "beta")

coda_diag(line, parameters = c("alpha", "beta"))

coda_diag(line, parameters = grep("sig", varnames(line), value = TRUE))

coda_diag(line, parameters = grep("a", varnames(line), value = TRUE))

coda_diag(line, parameters = c("alpha", grep("sig", varnames(line),
         value = TRUE)))
}
\author{
Michael Malick
}
\seealso{
\code{\link{traceplot}} \code{\link{density}}
         \code{\link{gelman.plot}} \code{\link{cumuplot}} \code{\link{hist}}
}

