% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_dotplot.R
\name{coda_dotplot}
\alias{coda_dotplot}
\title{Create a dotplot of median parameter values with 95\% credibility
       intervals}
\usage{
coda_dotplot(coda.object, parameters = NULL)
}
\arguments{
\item{coda.object}{An mcmc.list object}

\item{parameters}{character vector of parameter names to include in graphic. If none are
supplied all monitored parameters are included.}
}
\value{
A graphics device
}
\description{
\code{coda_dotplot} creates a single dotplot of the specified parameters
     (or all parameters if none are specified) with the dots giving the
     median value over all mcmc chains and horizontal bars giving the 80\%
     credibility interval (thick bars) and the 95\% credibility intervals
     (thin bars). This type of plot is also known as a catepillar plot and
     provides a way to easily summarize many parameters in a hierarchical
     model.
}
\examples{
library(coda)
data(line)

coda_dotplot(line)

coda_dotplot(line, parameters = "alpha")

coda_dotplot(line, parameters = c("alpha", "beta"))

coda_dotplot(line, parameters = grep("sig", varnames(line), value = TRUE))

coda_dotplot(line, parameters = grep("a", varnames(line), value = TRUE))

coda_dotplot(line, parameters = c("alpha", grep("sig", varnames(line),
             value = TRUE)))

}
\seealso{
\code{\link{xyplot}}
}
\author{
Michael Malick
}
