% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_grep.R
\name{coda_grep}
\alias{coda_grep}
\title{Quickly subset an mcmc.list object using grep}
\usage{
coda_grep(coda.object, pattern, return.matrix = FALSE, ...)
}
\arguments{
\item{coda.object}{an mcmc.list object}

\item{pattern}{character string giving the pattern to search for}

\item{return.matrix}{should the mcmc.list object be returned as a single matrix. If TRUE,
each element of the mcmc.list is binded together using rbind.}

\item{\dots}{additional arguments for \code{\link{grep}}}
}
\value{
An mcmc, mcmc.list or matrix
}
\description{
\code{coda_grep} takes as input an mcmc.list object and uses
     grep to find the parameters that match the specified pattern.
}
\examples{
library(coda)
data(line)

x <- coda_grep(line, "a")
head(x[[1]])

x <- coda_grep(line, "beta")
head(x[[1]])

x <- coda_grep(line, "sig")
head(x[[1]])

x <- coda_grep(line, "a", return.matrix = TRUE)
class(x)
dim(x)
head(x)

}
\seealso{
\code{\link{grep}}
}
\author{
Michael Malick
}
