% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coda_hist.R
\name{coda_hist}
\alias{coda_hist}
\title{Create a histogram of marginal posterior distributions}
\usage{
coda_hist(coda.object, parameters = NULL)
}
\arguments{
\item{coda.object}{An mcmc.list object}

\item{parameters}{character vector of parameter names to include in graphic. If none are
     supplied all monitored parameters are included.}
}
\value{
A graphics device
}
\description{
\code{coda_hist} uses the \code{hist} function inbase R
     to plot histograms of the marginal posterior
     distribution for each parameter specified by the parameters argument.
     In addition, a density overlay, rug of mcmc sample values, median value
     (bold vertical dash), and 95\% credibility interval (bold horizontal
     line) are plotted. A single histogram is printed per graphics device.
}
\examples{
library(coda)
data(line)

coda_hist(line)

coda_hist(line, parameters = "alpha")

coda_hist(line, parameters = c("alpha", "beta"))

coda_hist(line, parameters = grep("sig", varnames(line), value = TRUE))

coda_hist(line, parameters = grep("a", varnames(line), value = TRUE))

coda_hist(line, parameters = c("alpha", grep("sig", varnames(line),
               value = TRUE)))
}
\author{
Michael Malick
}
\seealso{
\code{\link{hist}}
}

