% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_neff.R
\name{coda_neff}
\alias{coda_neff}
\title{Create effective sample size diagnostic graphic}
\usage{
coda_neff(coda.object, plot = TRUE, thresh = 0.5, cex.labels = 0.5)
}
\arguments{
\item{coda.object}{an mcmc.list object}

\item{plot}{logical, should a plot be printed}

\item{thresh}{threshold used to flag parameters with low effective sample sizes, as a
percentage of total MCMC draws}

\item{cex.labels}{scaling for parameters plotted by name}
}
\value{
Invisibly returns a vector of the effective sample size for each
     parameter sorted from lowest (first) to highest (last). Optionally
     prints a plot of the effective sample sizes.
}
\description{
\code{coda_neff} creates a graphic of the effective sample size for each
     parameter in the input object. The left-hand y-axis is the total
     effective size and the right-hand y-axis is the effective size
     scaled by the total number of MCMC draws.
}
\examples{
library(coda)
data(line)

coda_neff(line)
}
\seealso{
\code{\link{effectiveSize}}
}
\author{
Michael Malick
}
