% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_rhat.R
\name{coda_rhat}
\alias{coda_rhat}
\title{Create Rhat diagnostic graphic}
\usage{
coda_rhat(rhat, thresh = 1.01, cex.labels = 0.5)
}
\arguments{
\item{rhat}{named vector of Rhat values}

\item{thresh}{threshold used to flag low Rhat values}

\item{cex.labels}{scaling for parameters plotted by name}
}
\value{
Invisibly returns a vector of the Rhat values for each
     parameter sorted from highest (first) to lowest (last)
     and prints a plot of the effective sample sizes.
}
\description{
\code{coda_rhat} creates a graphic of the potential scale reduction
     factor (Rhat)
}
\examples{
library(coda)
data(line)

gd <- coda::gelman.diag(line)
rhat <- gd$psrf[ , 1]
coda_rhat(rhat)
}
\seealso{
\code{\link{gelman.diag}}
}
\author{
Michael Malick
}
