% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_table.R
\name{coda_table}
\alias{coda_table}
\title{Create a summary table of an mcmc.list object}
\usage{
coda_table(coda.object, parameters = NULL)
}
\arguments{
\item{coda.object}{an mcmc.list object}

\item{parameters}{character vector of parameter names to include in table. If none are
supplied all monitored parameters are included.}
}
\value{
A dataframe
}
\description{
\code{coda_table} takes as input an mcmc.list object from the coda
     package and outputs a dataframe giving the mean, standard
     deviation, and median for each parameter estimate along with quantiles
     for 95\% and 80\% credibility intervals.
}
\examples{
library(coda)
data(line)

coda_table(line)

coda_table(line, parameters = "alpha")

coda_table(line, parameters = c("alpha", "beta"))

coda_table(line, parameters = grep("sig", coda::varnames(line), value = TRUE))

coda_table(line, parameters = grep("a", coda::varnames(line), value = TRUE))

coda_table(line, parameters = c("alpha", grep("sig", coda::varnames(line),
         value = TRUE)))

}
\seealso{
\code{\link{coda_df}}
}
\author{
Michael Malick
}
