% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_density.R
\name{col_density}
\alias{col_density}
\title{Plots non-parametric density overlays for each column in a matrix}
\usage{
col_density(x, xlim = NULL, xlab = "", main = "", plot.it = TRUE,
  ...)
}
\arguments{
\item{x}{a matrix}

\item{xlim}{vector of length 2 giving min and max for x-axis}

\item{xlab}{x-axis label}

\item{main}{main title for graphic}

\item{plot.it}{logical, should a plot be created}

\item{\dots}{additional arguments passed to \code{density}}
}
\value{
Invisibly returns a list with x and y components for the plotted density
     curves.
}
\description{
Calculate and plot the non-parametric density for the data in each
     column of a matrix. All densities are plotted on the same graphic.
}
\examples{
 mat <- matrix(rnorm(100000), ncol = 100)
 col_density(mat)
 col_density(mat, main = "test")
 col_density(mat, xlim = c(-4, 3))
 xx <- col_density(mat, xlim = c(-4, 3))
 class(xx)
}
\seealso{
\code{\link{density}}
}
\author{
Michael Malick
}
