% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coda_histogram.R
\name{coda_histogram}
\alias{coda_histogram}
\title{Create a histogram of marginal posterior distributions}
\usage{
coda_histogram(coda.object, parameters = NULL)
}
\arguments{
\item{coda.object}{An mcmc.list object}

\item{parameters}{character vector of parameter names to include in graphic. If none are
     supplied all monitored parameters are included.}
}
\value{
A graphics device
}
\description{
\code{coda_histogram} plots a histogram of the marginal posterior
     distribution for each parameter specified by the parameters argument.
     In addition, a density overlay, rug of mcmc sample values, median value
     (bold vertical dash), and 95\% credibility interval (bold horizontal
     line) are plotted. The function uses lattice to panel the histograms for
     each parameter.
}
\examples{
library(coda)
data(line)

coda_histogram(line)

coda_histogram(line, parameters = "alpha")

coda_histogram(line, parameters = c("alpha", "beta"))

coda_histogram(line, parameters = grep("sig", varnames(line), value = TRUE))

coda_histogram(line, parameters = grep("a", varnames(line), value = TRUE))

coda_histogram(line, parameters = c("alpha", grep("sig", varnames(line),
               value = TRUE)))
}
\author{
Michael Malick
}
\seealso{
\code{\link{histogram}}
}

