% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_neff.R
\name{coda_neff}
\alias{coda_neff}
\title{Create effective sample size diagnostic graphic}
\usage{
coda_neff(neff, samps, thresh = 0.5, cex.labels = 0.5)
}
\arguments{
\item{neff}{named vector of Neff values}

\item{samps}{total number of MCMC draws (N chains * N draws)}

\item{thresh}{threshold used to flag parameters with low effective sample sizes, as a
percentage of total MCMC draws}

\item{cex.labels}{scaling for parameters plotted by name}
}
\value{
Invisibly returns a vector of the effective sample size for each
     parameter sorted from lowest (first) to highest (last) and
     prints a plot of the effective sample sizes.
}
\description{
\code{coda_neff} creates a graphic of the effective sample size.
     The left-hand y-axis is the total effective size and the right-hand
     y-axis is the effective size scaled by the total number of MCMC draws.
}
\examples{
library(coda)
data(line)

neff  <- coda::effectiveSize(line)
samps <- coda::niter(line) * coda::nchain(line)
coda_neff(neff, samps)
}
\seealso{
\code{\link{effectiveSize}}
}
\author{
Michael Malick
}
