% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_rhat.R
\name{coda_rhat}
\alias{coda_rhat}
\title{Create Rhat diagnostic graphic}
\usage{
coda_rhat(coda.object, plot = TRUE, thresh = 1.01, cex.labels = 0.5)
}
\arguments{
\item{coda.object}{an mcmc.list object}

\item{plot}{logical, should a plot be printed}

\item{thresh}{threshold used to flag low Rhat values}

\item{cex.labels}{scaling for parameters plotted by name}
}
\value{
Invisibly returns a vector of the Rhat values for each
     parameter sorted from highest (first) to lowest (last). Optionally
     prints a plot of the effective sample sizes.
}
\description{
\code{coda_rhat} creates a graphic of the potential scale reduction
     factor (Rhat) for each parameter in the input
}
\examples{
library(coda)
data(line)

coda_rhat(line)
}
\seealso{
\code{\link{gelman.diag}}
}
\author{
Michael Malick
}
