% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colDensity.R
\name{colDensity}
\alias{colDensity}
\title{Plots non-parametric density overlays for each column in a matrix}
\usage{
colDensity(x, xlim = NULL, xlab = "", main = "", ...)
}
\arguments{
\item{x}{a matrix}

\item{xlim}{vector of length 2 giving min and max for x-axis}

\item{xlab}{x-axis label}

\item{main}{main title for graphic}

\item{\ldots}{additional argurments passed to \code{density}}
}
\value{
Invisibly returns a list with x and y components for the plotted density
     curves.
}
\description{
Calculate and plot the non-parametric density for the data in each
     column of a matrix. All densities are plotted on the same graphic.
}
\examples{
mat <- matrix(rnorm(100000), ncol = 100)
colDensity(mat)
colDensity(mat, main = "test")
colDensity(mat, xlim = c(-4, 3))
xx <- colDensity(mat, xlim = c(-4, 3))
class(xx)
}
\author{
Michael Malick
}
\seealso{
\code{\link{density}}
}

